+++ app/models/user.rb
@@ -2,6 +2,7 @@ class User < ActiveRecord::Base
   has_one :profile, :dependent => :destroy
   belongs_to :gender
   has_many :posts, :dependent => :destroy
+  has_many :comments, :dependent => :destroy
 
   has_secure_password