+++ db/migrate/20151116061618_create_comments.rb
@@ -1,0 +1,15 @@
+class CreateComments < ActiveRecord::Migration
+  def change
+    create_table :comments do |t|
+      t.text :body
+      t.integer :user_id
+      t.integer :commentable_id
+      t.string :commentable_type
+
+      t.index :user_id
+      t.index :commentable_id
+
+      t.timestamps null: false
+    end
+  end
+end